[refactor] 스킬 받아오기가 더 매끄럽도록 변경

+++ BE/src/main/java/cuketmon/skill/service/SkillService.java
@@ -69,6 +69,11 @@ public class SkillService {
     // ㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡㅡ
     // 전체 스킬 저장 (919개)
     public void fetchAndSaveAllSkills() {
+        // 919번 스킬이 존재하면 전체 저장 생략
+        if (skillRepository.existsById(TOTAL_SKILL)) {
+            return;
+        }
+
         skillRepository.deleteAll();
         for (int i = 1; i <= TOTAL_SKILL; i++) {
             fetchAndSaveSkill(i);
